added comments
realized that this might not exactly be what i was going for. I was trying to create a radom page that displayed 3 of the venues randomly upon clicking the discover button. I however made a way for the user to click on a button and 3 random venues would be selected. Needs further testing.